+++ Intro_to_objective-c/Person.h
@@ -14,7 +14,7 @@
 
 @property(strong, nonatomic) NSString *firstName;
 @property(strong,nonatomic) NSString *lastName;
-@property int age;
+@property NSInteger age;
 
 
 //name accessor methods
